// models/response/proposedMobilityProgrammeResponse.ts
export interface ProposedMobilityProgrammeResponse {
    id: number;
    plannedStartingDateOfMobility: string;
    plannedEndDateOfMobility: string;
    receivingInstitutionCourseCatalogueLink: string;
    language_id: number;
    languageLevel_id: number;
    provisionsLinkIfEducationUnsuccessful: string;
}